Delete ComplaintService. Add all validations back to servlet.
