+++ src/main/java/com/cops/ntsf/controller/ComplaintServlet.java
@@ -1,7 +1,6 @@
 package com.cops.ntsf.controller;
 
 import com.cops.ntsf.model.Complaint;
-import com.cops.ntsf.service.ComplaintService;
 import com.google.gson.Gson;
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -44,10 +43,7 @@ public class ComplaintServlet extends HttpServlet {
 
         Complaint complaint = null;
 
-        // Validations done in complaintService
-        ComplaintService complaintService = new ComplaintService();
-
-        if (complaintService.checkValidations(title, description)) {
+        if (checkValidations(title, description)) {
             complaint = new Complaint(user_id, title, description, complaint_no);
             complaint.complaintAdded();
         }
@@ -95,5 +91,47 @@ public class ComplaintServlet extends HttpServlet {
 
         out.write(jsonObject.toString());
         out.close();
+    }
+
+    /*
+    @ Check Validations function
+    * */
+    public boolean checkValidations(String title, String description) {
+
+        boolean flagTitle = false; // flag = true means title validation is passed
+        boolean flagDescription = false; // flag = true means description validation is passed
+
+        /*
+        @ Validate title
+        * */
+        if (title.trim().equals("")) {
+            System.out.println("Title is empty");
+        } else if (title.trim().length() < 3) {
+            System.out.println("Title is too short");
+        } else if (title.trim().length() > 50) {
+            System.out.println("Title is too long");
+        } else {
+            System.out.println("Title is valid");
+            flagTitle = true;
+        }
+
+        /*
+        @ Validate description
+        * */
+        if (description.trim().equals("")) {
+            System.out.println("Description is empty");
+        } else if (description.trim().length() < 10) {
+            System.out.println("Description is too short");
+        } else if (description.trim().length() > 500) {
+            System.out.println("Description is too long");
+        } else {
+            System.out.println("Description is valid");
+            flagDescription = true;
+        }
+
+        // Simplify 'if else'
+        boolean flag;
+        flag = flagTitle && flagDescription;
+        return flag;
     }
 }
